# Clean HTML stuff, add Timeline.WriteHTMLTo