Fixes #32

`Timeline.WriteHTMLTo` takes a `io.Writer` so that it can directly
be used in various cases, for example directly passing an
http.ResponseWriter. Using a writer, many allocations are removed
by directly formatting the output with `fmt.Fprint`, remove the
`bytes.Buffer. as well.

Use a `text/template` for HTML generation. An `html/template` are
very similar but escapes the strings, and it was not playing
well the JSON marshaling of graph labels, also it is not based on
user input so no risk of code injection.

`genGraphHTML` also takes a writer to eliminate more copies.

Modify `Timeline.WriteHTML` to keep backward compatibility but
make it use WriteHTMLTo under the hood.

Add `BenchmarkWriteHTMLTo`. In the following benchmarks, `old` is
a version of `WriteHTMLTo` equivalent to original `WriteHTML`
code, present in current master, whereas `new` is this commit.

    name           old time/op    new time/op    delta
    WriteHTMLTo-8    25.4µs ±12%     7.0µs ± 3%  -72.49%  (p=0.008 n=5+5)

    name           old alloc/op   new alloc/op   delta
    WriteHTMLTo-8     161kB ± 0%       1kB ± 0%  -99.37%  (p=0.008 n=5+5)

    name           old allocs/op  new allocs/op  delta
    WriteHTMLTo-8      56.0 ± 0%      50.0 ± 0%  -10.71%  (p=0.008 n=5+5)